move fw data bridge mini lcm tests into folder
