+++ backend/FwLite/FwDataMiniLcmBridge.Tests/MiniLcmTests/CreateEntryTests.cs
@@ -1,8 +1,6 @@
-﻿using FwDataMiniLcmBridge.Api;
-using FwDataMiniLcmBridge.Tests.Fixtures;
-using MiniLcm.Models;
+﻿using FwDataMiniLcmBridge.Tests.Fixtures;
 
-namespace FwDataMiniLcmBridge.Tests;
+namespace FwDataMiniLcmBridge.Tests.MiniLcmTests;
 
 [Collection(ProjectLoaderFixture.Name)]
 public class CreateEntryTests(ProjectLoaderFixture fixture) : CreateEntryTestsBase
